Prep/KPI-API/requirements.txt
fastapi==0.95.2
pydantic==1.10.13
mangum==0.17.0
python-dotenv